Imported from bmake-20120518.tar.gz.

+++ meta.c
@@ -774,8 +774,8 @@ fgetLine(char **bufp, size_t *szp, int o, FILE *fp)
 	    if (newsz <= bufsz)
 		newsz = ROUNDUP(fs.st_size, BUFSIZ);
 	    if (DEBUG(META)) 
-		fprintf(debug_file, "growing buffer %zu -> %zu\n",
-			bufsz, newsz);
+		fprintf(debug_file, "growing buffer %u -> %u\n",
+			(unsigned)bufsz, (unsigned)newsz);
 	    p = bmake_realloc(buf, newsz);
 	    if (p) {
 		*bufp = buf = p;
